# [addons] add define ATTRIBUTE_FORCEINLINE on addon header