@pkerling 2019-01-18
Curious that this would make a discernible difference - what kind of performance gain do you get e.g. in percent?