@AlwinEsch 2019-01-18
Is only a little bit, but in the last addon I had to clean up the code and remove the whole "static" and "export" integrate everything into classes, which can be extremely many calls in the stream.

Just thought to make it as efective as possible to stay close to the previous "C" in performance.

In static included class:
* Without inline: 0.00017 s
* With inline: 0.00015 s

In pointer included class:
* Without inline: 0.00014 s
* With inline: 0.00010 s

Funnily enough, the class is faster with a pointer call than static with "."

Here the quick test code I made now:
```cpp
#include <chrono>
#include <iostream>

class CAAA
{
public:
  CAAA() = default;

  inline __attribute__((__always_inline__)) bool GetA() { return aaa; }

private:
  bool aaa = true;
};

int main(int argc, char **argv)
{
  CAAA aaab;
  CAAA* aaab2 = CAAA;

    auto start = std::chrono::high_resolution_clock::now();

    for (unsigned int i = 0; i < 100000; i++)
    {
      bool ret;
      if (aaab.GetA())
        ret = true;
      else
        ret = false;
    }

    auto finish = std::chrono::high_resolution_clock::now();
    std::chrono::duration<double> elapsed = finish - start;
    std::cout << "Elapsed false time: " << elapsed.count() << " s\n";
}
```